Create sql
Added the file where i had related all the queries to create a database of a worker management system  application. And one thing is that the queries are written in such a way and sequence that the whole queries can run at once by just copy paste the whole code of the file.